+++ src/content/config.ts
@@ -39,6 +39,18 @@ const states = defineCollection({
   }),
 });
 
-const blog = defineCollection({ type: "content" });
+const blog = defineCollection({
+  type: "content",
+  schema: z.object({
+    title: z.string(),
+    description: z.string(),
+    pubDate: z.coerce.date(),
+    updatedDate: z.coerce.date().optional(),
+    author: z.string().optional(),
+    image: z.string().optional(),
+    imageAlt: z.string().optional(),
+    tags: z.array(z.string()).optional(),
+  }),
+});
 
 export const collections = { installers, states, blog };